Adjust heightForString div offscreen.
The div used in SC.heightForString is left in the DOM. If the most recent
height measured was taller than the window, then in cases where the body
element's overflow was set to auto, the scrollable height would match
this div's height.

To avoid this, after we determine the height, set the div's top to -height.