Clear selection if collection view delegate removes all indexes.
The behavior of SC.CollectionView#select was different depending on
whether an empty/null index list was passed in or returned by the delegate.

In the case of passing in a empty/null list, the selection would be
cleared.

In the case of passing in a non-empty list that the delegate's
collectionViewShouldSelectIndexes later clears, the function would
immediately return, leaving any previous selection intact.

If the delegate clear the index list, it will be set to null so that the
selection will also be cleared in this case.