Move the notification of property chains into _notifyPropertyObservers.
Since the notification was happening in propertyDidChange, before the
cached value of the property was cleared, if a dependent property chain
used the value of the property that had changed, it would get the old
cached value.

Moving the notification into _notifyPropertyObservers ensures it happens
after caches are cleared, and also makes the behavior consistent with all
other notifications (i.e. obeying _kvo_changeLevel and
isObservingSuspended).