Skip branch observer invalidation on parents.
When the children of a SC.TreeItemObserver would change, all branch
observers all the way of the tree would be destroyed. Since only the
children at the lowest level changed, only those branch observers need to
be destroyed.

This avoids a lot of extra object creation, speeding up large trees.