Further fixes to firstObject/lastObject invalidation.
The check in _scac_arrayContentDidChange to determine if lastObject should
be invalidated was slightly wrong and would invalidate too often.

The check in enumerableContentDidChange had the same problem, and if the
parameters were not passed in, firstObject and lastObject would never be
invalidated. While we know in this case that firstObject and lastObject
are not intended to be observed, since they are cacheable properties, the
cache would not be invalidated, so it was not safe to use get() on these
properties.

Also added unit tests to make sure firstObject and lastObject are
invalidated when expected (and not more often).