Cleanup SC.SegmentedView overflow menu.
There were two problems here. When the overflow menu was removed, it was
never being destroyed, causing a memory leak. Second, when the segmented
view was destroyed, the overflow menu was not being removed, so you would
end up with a dangling menu in this case, causing crashes if you tried to
interact with it.