Improve Touch Scrolling
Use capturing to prevent side effects when touch scrolling, such as
accidentally changing the selection of a scrolled list. Added the ability
to have touchesDragged not trigger a scroll if a view doesn't want
scrolling. Will be used to allow touch drag & drop.

Added touch support to scrollers

Removed SC.device.mouseLocation which was only used by the scrollers and
was incorrect for touch. SC.ScrollerView now tracks the coordinates
internally.