Support Touch Drag & Drop
Made changes to collection view to properly support selectOnMouseDown so
you must mouse down & up on a row in order for it to be selected and so
that when selecting on mouse down dragging a selected row drags all
selected rows.

Made changes to collectionViewShouldBeginDrag to receive the correct
indices of the proposed drag and to receive the event that triggered the
drag operation, allowing for touch specific logic.

For browsers that send Touch Events by default prevent auto scrolling of
scroll views when dragging to avoid cases where the target of the
touchstart get removed from the DOM due to incremental rendering which
breaks the drag & drop.

For list items handle the drop target CSS class via a classNameBinding
instead of via a displayPropery to prevent re-rendering during drag & drop
which can also break drag & drop for Touch Events.